pyproject.toml
[project]
name = "fakenews-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "contractions>=0.1.73",
    "en-core-web-sm",
    "fastparquet>=2024.11.0",
    "huggingface-hub[cli]>=0.27.1",
    "ipykernel>=6.29.5",
    "joblib>=1.4.2",
    "lightgbm>=4.5.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "numpy==1.*",
    "optuna>=4.1.0",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "plotly>=5.24.1",
    "protobuf==3.20.*",
    "psrecord>=1.4",
    "psutils>=3.3.9",
    "pyarrow>=18.1.0",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.2",
    "shap>=0.46.0",
    "spacy>=3.8.3",
    "streamlit>=1.41.1",
    "tensorflow==2.9.0",
    "torch>=2.5.1",
    "transformers>=4.48.0",
    "wordcloud>=1.9.4",
]

[tool.scripts]
download-spacy-model = "python -m spacy download en_core_web_sm"

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
